completed all loops problems

+++ Codeforces/100xSchoolQuestions/loops/h_sum_of_first_n_natural_numbers.cpp
@@ -2,9 +2,11 @@
 using namespace std;
 
 int main(){
-	int n;
+//NOTE: USING INT DOESN'T SATISFY EDGE CASES WHERE LARGE NUMBERS ARE ENTERED
+//      THAT'S WHY LONG LONG IS BEING USED
+	long long n;
 	cin >> n;
-	int sum{0};
+	long long sum{0};
 	for(int i = 1;i<=n;i++){
 		sum += i;
 	}
